Finished moving tests around, but still a number of test failures to fix.
