+++ glue/viewers/histogram_new/qt/tests/test_viewer_widget.py
@@ -333,3 +333,12 @@ class TestHistogramViewer(object):
     #     assert self.viewer.component is test
 
     # TODO: Check for extraneous draw events
+
+    # def test_nbin_override_persists_over_attribute_change(self):
+    #     # regression test for #398
+    #     self.collect.append(self.data)
+    #     self.client.add_layer(self.data)
+    #     self.client.set_component(self.data.id['x'])
+    #     self.client.nbins = 7
+    #     self.client.set_component(self.data.id['y'])
+    #     assert self.client.nbins == 7
